requirements/tests.txt
lxml==3.4.4
namedlist==1.7
py==1.4.28
pytz==2015.4
pytest==2.8.7
pytest-allure-adaptor==1.6.5
pytest-cov==2.2.1
pytest-django==2.8.0
